UI Layout has installer stages on left, main installer pane on right.
The 'Screenshots' directory shows screenshots for Apple's SF Symbols
installer, which has more installer steps than this project will
have. This installer will only install software for all users of
the computer it's installed on, and will always install to the
volume the computer is booted from.

Clicking the 'Continue' button repeatedly will drive the UI
through the Introduction, Licence, Installation and Summary
screens. No actual installation is performed.

Introduction - presents the user with a description of what
the installer is for, and what will be installed.

Licence - presents an end-user agreement for the user to read.
The 'Print' and 'Save' buttons do nothing in this commit,
and the user is not prompted to accept or reject the licence
before continuing in this commit.

Installation - show a list of actions that the installer is
performing, with progress indicators. In this commit, there are
no actions performed, and advancing past this screen is automatic
once all installation actions have completed successfully (in which
case the installation is considered to have succeeded), or some
installation step has failed (in which case the installation has failed).

Summary - Displays a short explanation of what the installer has done.
The possible outcomes are:
Licence rejected - the software will not be installed because the user
rejected the licence.
Success - the user accepted the licence and all installation actions
completed successfully. If the software was installed beforehand, then
it has been replaced with the version in the installer.
Failure with rollback - an earlier installation of the software was
detected, but not all installation steps completed successfully.
The installer has put things back to the way they were before it was run.