License is read from an HTML file, which is displayed in a web view.
Rejecting the license takes the user to the summary page, where
the user is told that nothing has been (or can) be installed
unless the user agrees to the license.

Accepting the license takes the user to the installation screen,
where a set of installation steps are shown. The installer fakes
the experience of installation by pausing for five seconds, then
treating the installation as successful.

The next step is to add a background service target to the app, which
will run as root and enable the UI to drive the installation of files
into directories that only root has write access to.

This will be used to copy a set of files embedded in the installer app
to their destinations.