AdjustProjectSettings can be undone with RestoreProjectSettings
The script ThereAndBack is meant to take a clean checkout of
the repository and customize the bundle identifiers for the
installer application and the helper application and set the
development team used to codesign both using AdjustprojectSettings,
then undo the customization using RestoreProjectSettings.

macInstaller/HELPERNAME/Info.plist and macInstaller/macInstaller/Info.plist
needed to have the embedded strings changed to use " instead
of &quot; as delimiters, because plutil can write strings with "
in them, but trying to write '&quot;' caused various XML escapes
to be applied to the string before it was written, making the final result
not match what the script source suggested it would be. This looked
like a rich source of silent bugs.