Installation consists of a set of actions, carried out in order.
During the automated install process, the background service
will be invoked to perform all the individual steps which,
put together, result in the software we want to install be installed
or, if an error occurs, the previously installed software is
restored to the state it was before the installer was launched.

The actions are:
- the service is stopped (and verified to not be running)
- the files listed in the file 'PayloadMetadata.plist' are installed
into their destination directories, and their owner/group and permissions
are set according to PayloadMetadata.plist.
To avoid a bad install, the installer will enforce the rules required
by macOS: files in /Library must be owned by root, and executables must
not be writeable.
- the newly-installed service is started (and verified to be running)
- any files that were updated during the install will have been saved
in a 'restore' directory. If all the previous steps have completed
successfully, then they are removed.
- If any of the steps failed to complete successfully, then the saved
files are copied back to their original locations (undoing the install),
and the previous version of the service is restarted.