Add Installer actions (SwiftUI)
The installation takes place by executing steps, one
after the other. This version of the installer can install
a sample system-level service 'com.greenkitty.product' in the Resources/
folder.

In order to change the payload, run AdjustProjectSettings to replace the
placeholder bundle identifier for the installer executable with your own,
then run SetupPayloadMetadata.zsh (suitably edited) to replace
Resources/PayloadMetadata.plist with the values needed for your payload.
Payload/com.greenkitty.product and Payload/com.greenkitty.product.plist
can be discarded, since they are placeholders for your payload.

If you build and run the project with the placeholder payload, a system-
level service com.greenkitty.product will be installed which logs to
/var/log/product.log. What is logs is the current logged-in user each time
the service starts or the logged-in user changes.

If you install the service
and log out of the Finder (leaving the login window up), and then log
back in, you should see lines in the log showing that no user was logged in
between entries saying that a user was logged in - this shows that the service
is running as a system service not bound to any user's session.