Add 'Install Scenarios'
If the metadata specifies more than one install scope (out of 'System',
'AllUsers' and 'User'), then the 'choose an installation scope' screen
is shown and the user will be able to choose one.

If only one install scope is specified, then the scope selection screen is skipped.

To do:
- test with more than one install scope available.
- service start/stop using launchctl cannot be run as root unless the scope
is 'System'.